src/modules/lab_test_management/requirements.txt
pymongo
python-dotenv
streamlit
streamlit-option-menu
pandas
matplotlib
